refactor(#3983): move the strict [[Set]] builder into its own subsystem module
The #3102 LOC-regrowth ratchet rejected the inline shape (object-runtime.ts
+94 over its 8,322 budget). Taking a `loc-budget-allow:` grant would have been
the wrong answer -- the gate is asking for exactly the right thing here, since
the builder is self-contained and has no reason to live in the god-file.

Extract it to src/codegen/object-runtime-strict-set.ts (which also now carries
the full rationale: why it layers over `__reflect_set` instead of re-deriving
the flag walk, and why the non-`$Object` receiver short-circuit is load-bearing
rather than defensive). object-runtime.ts keeps a 3-line call site. The gate now
passes with NO allowance; object-runtime.ts is 8321, one under its budget.

Because this refactors already-measured code, the AFTER arm was re-run from
scratch against the same BEFORE arm: 244 pass / 337 -- identical, with an
identical per-file diff (fam +23/-0 plus the same floored 24th flip, and the
same 5 already-de-bunked control flakes). Behaviour-identical.
tests/issue-3983.test.ts still 14/14.

Also pre-ran every cheap `quality` gate rather than only the ones the last
failure named -- `quality` is fail-fast under `bash -e`, so an early red means
the later gates emit no verdict at all. 13/14 pass. The 14th,
`check:issue-ids:against-main`, fails ONLY locally: `origin` in this checkout is
the FORK, whose stale `main` carries different issue files at #3973/#3974/#3983.
Against the real CI base it is clean --
`GATE_BASE=upstream/main node scripts/check-issue-ids.mjs --against-main` passes,
`--against-open-prs` passes, and upstream/main has no #3983 file at all.


✓
